move dot-based GEMM out of TPL CUBLAS
